feat: initial commit, exporting op codes to evm-module-api
