+++ hedera-node/hedera-evm-api/src/main/java/com/hedera/services/evm/contracts/operations/HederaExtCodeHashOperation.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hedera.services.contracts.operation;
+package com.hedera.services.evm.contracts.operations;
 
 /*
  * -
@@ -37,27 +37,32 @@ package com.hedera.services.contracts.operation;
  *
  */
 
-import static org.hyperledger.besu.evm.internal.Words.clampedToLong;
-
-import java.util.function.BiPredicate;
+import org.apache.tuweni.units.bigints.UInt256;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.evm.EVM;
+import org.hyperledger.besu.evm.frame.ExceptionalHaltReason;
 import org.hyperledger.besu.evm.frame.MessageFrame;
 import org.hyperledger.besu.evm.gascalculator.GasCalculator;
-import org.hyperledger.besu.evm.operation.ExtCodeCopyOperation;
+import org.hyperledger.besu.evm.internal.FixedStack;
+import org.hyperledger.besu.evm.internal.Words;
+import org.hyperledger.besu.evm.operation.ExtCodeHashOperation;
+
+import java.util.Optional;
+import java.util.OptionalLong;
+import java.util.function.BiPredicate;
 
 /**
- * Hedera adapted version of the {@link ExtCodeCopyOperation}.
+ * Hedera adapted version of the {@link ExtCodeHashOperation}.
  *
  * <p>Performs an existence check on the requested {@link Address} Halts the execution of the EVM
  * transaction with {@link HederaExceptionalHaltReason#INVALID_SOLIDITY_ADDRESS} if the account does
  * not exist or it is deleted.
  */
-public class HederaExtCodeCopyOperation extends ExtCodeCopyOperation {
+public class HederaExtCodeHashOperation extends ExtCodeHashOperation {
 
     private final BiPredicate<Address, MessageFrame> addressValidator;
 
-    public HederaExtCodeCopyOperation(
+    public HederaExtCodeHashOperation(
             GasCalculator gasCalculator, BiPredicate<Address, MessageFrame> addressValidator) {
         super(gasCalculator);
         this.addressValidator = addressValidator;
@@ -65,14 +70,33 @@ public class HederaExtCodeCopyOperation extends ExtCodeCopyOperation {
 
     @Override
     public OperationResult execute(MessageFrame frame, EVM evm) {
-        final long memOffset = clampedToLong(frame.getStackItem(1));
-        final long numBytes = clampedToLong(frame.getStackItem(3));
+        try {
+            final Address address = Words.toAddress(frame.popStackItem());
+            if (!addressValidator.test(address, frame)) {
+                return new OperationResult(
+                        OptionalLong.of(cost(true)),
+                        Optional.of(HederaExceptionalHaltReason.INVALID_SOLIDITY_ADDRESS));
+            }
+            final var account = frame.getWorldUpdater().get(address);
+            boolean accountIsWarm =
+                    frame.warmUpAddress(address) || this.gasCalculator().isPrecompile(address);
+            OptionalLong optionalCost = OptionalLong.of(this.cost(accountIsWarm));
+            if (frame.getRemainingGas() < optionalCost.getAsLong()) {
+                return new OperationResult(
+                        optionalCost, Optional.of(ExceptionalHaltReason.INSUFFICIENT_GAS));
+            } else {
+                if (!account.isEmpty()) {
+                    frame.pushStackItem(UInt256.fromBytes(account.getCodeHash()));
+                } else {
+                    frame.pushStackItem(UInt256.ZERO);
+                }
 
-        return HederaOperationUtil.addressCheckExecution(
-                frame,
-                () -> frame.getStackItem(0),
-                () -> cost(frame, memOffset, numBytes, true),
-                () -> super.execute(frame, evm),
-                addressValidator);
+                return new OperationResult(optionalCost, Optional.empty());
+            }
+        } catch (final FixedStack.UnderflowException ufe) {
+            return new OperationResult(
+                    OptionalLong.of(cost(true)),
+                    Optional.of(ExceptionalHaltReason.INSUFFICIENT_STACK_ITEMS));
+        }
     }
 }
