+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/ContractsV_0_30Module.java
@@ -21,16 +21,16 @@ import static org.hyperledger.besu.evm.operation.SStoreOperation.FRONTIER_MINIMU
 import com.hedera.services.context.TransactionContext;
 import com.hedera.services.context.properties.GlobalDynamicProperties;
 import com.hedera.services.contracts.ContractsModule.V_0_30;
-import com.hedera.services.contracts.operation.HederaBalanceOperation;
+import com.hedera.services.evm.contracts.operations.HederaBalanceOperation;
 import com.hedera.services.contracts.operation.HederaCallCodeOperation;
 import com.hedera.services.contracts.operation.HederaCallOperation;
 import com.hedera.services.contracts.operation.HederaChainIdOperation;
 import com.hedera.services.contracts.operation.HederaCreate2Operation;
 import com.hedera.services.contracts.operation.HederaCreateOperation;
-import com.hedera.services.contracts.operation.HederaDelegateCallOperation;
-import com.hedera.services.contracts.operation.HederaExtCodeCopyOperation;
-import com.hedera.services.contracts.operation.HederaExtCodeHashOperation;
-import com.hedera.services.contracts.operation.HederaExtCodeSizeOperation;
+import com.hedera.services.evm.contracts.operations.HederaDelegateCallOperation;
+import com.hedera.services.evm.contracts.operations.HederaExtCodeCopyOperation;
+import com.hedera.services.evm.contracts.operations.HederaExtCodeHashOperation;
+import com.hedera.services.evm.contracts.operations.HederaExtCodeSizeOperation;
 import com.hedera.services.contracts.operation.HederaLogOperation;
 import com.hedera.services.contracts.operation.HederaSLoadOperation;
 import com.hedera.services.contracts.operation.HederaSStoreOperation;
