+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/operation/HederaOperationUtil.java
@@ -38,8 +38,19 @@ package com.hedera.services.contracts.operation;
  */
 
 import com.hedera.services.contracts.sources.EvmSigsVerifier;
+import com.hedera.services.evm.contracts.operations.HederaEvmOperationsUtil;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.store.contracts.HederaStackedWorldStateUpdater;
 import com.hedera.services.store.contracts.HederaWorldState;
+import org.apache.commons.lang3.tuple.MutablePair;
+import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.bytes.Bytes32;
+import org.apache.tuweni.units.bigints.UInt256;
+import org.hyperledger.besu.datatypes.Address;
+import org.hyperledger.besu.evm.frame.MessageFrame;
+import org.hyperledger.besu.evm.operation.Operation;
+import org.hyperledger.besu.evm.precompile.PrecompiledContract;
+
 import java.util.Map;
 import java.util.Optional;
 import java.util.OptionalLong;
@@ -47,56 +58,11 @@ import java.util.TreeMap;
 import java.util.function.BiPredicate;
 import java.util.function.LongSupplier;
 import java.util.function.Supplier;
-import org.apache.commons.lang3.tuple.MutablePair;
-import org.apache.tuweni.bytes.Bytes;
-import org.apache.tuweni.bytes.Bytes32;
-import org.apache.tuweni.units.bigints.UInt256;
-import org.hyperledger.besu.datatypes.Address;
-import org.hyperledger.besu.evm.frame.ExceptionalHaltReason;
-import org.hyperledger.besu.evm.frame.MessageFrame;
-import org.hyperledger.besu.evm.internal.FixedStack;
-import org.hyperledger.besu.evm.internal.Words;
-import org.hyperledger.besu.evm.operation.Operation;
-import org.hyperledger.besu.evm.precompile.PrecompiledContract;
 
 /** Utility methods used by Hedera adapted {@link org.hyperledger.besu.evm.operation.Operation} */
 public final class HederaOperationUtil {
     private HederaOperationUtil() {
         throw new UnsupportedOperationException("Utility Class");
-    }
-
-    /**
-     * An extracted address check and execution of extended Hedera Operations. Halts the execution
-     * of the EVM transaction with {@link HederaExceptionalHaltReason#INVALID_SOLIDITY_ADDRESS} if
-     * the account does not exist, or it is deleted.
-     *
-     * @param frame The current message frame
-     * @param supplierAddressBytes Supplier for the address bytes
-     * @param supplierHaltGasCost Supplier for the gas cost
-     * @param supplierExecution Supplier with the execution
-     * @param addressValidator Address validator predicate
-     * @return The operation result of the execution
-     */
-    public static Operation.OperationResult addressCheckExecution(
-            MessageFrame frame,
-            Supplier<Bytes> supplierAddressBytes,
-            LongSupplier supplierHaltGasCost,
-            Supplier<Operation.OperationResult> supplierExecution,
-            BiPredicate<Address, MessageFrame> addressValidator) {
-        try {
-            final var address = Words.toAddress(supplierAddressBytes.get());
-            if (Boolean.FALSE.equals(addressValidator.test(address, frame))) {
-                return new Operation.OperationResult(
-                        OptionalLong.of(supplierHaltGasCost.getAsLong()),
-                        Optional.of(HederaExceptionalHaltReason.INVALID_SOLIDITY_ADDRESS));
-            }
-
-            return supplierExecution.get();
-        } catch (final FixedStack.UnderflowException ufe) {
-            return new Operation.OperationResult(
-                    OptionalLong.of(supplierHaltGasCost.getAsLong()),
-                    Optional.of(ExceptionalHaltReason.INSUFFICIENT_STACK_ITEMS));
-        }
     }
 
     /**
