+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/operation/HederaSelfDestructOperation.java
@@ -40,6 +40,7 @@ package com.hedera.services.contracts.operation;
 import static com.hedera.services.utils.EntityIdUtils.numOfMirror;
 
 import com.hedera.services.context.TransactionContext;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.store.contracts.HederaStackedWorldStateUpdater;
 import java.util.Optional;
 import java.util.OptionalLong;
