+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/operation/HederaStaticCallOperation.java
@@ -38,14 +38,17 @@ package com.hedera.services.contracts.operation;
  */
 
 import com.hedera.services.contracts.sources.EvmSigsVerifier;
-import java.util.Map;
-import java.util.function.BiPredicate;
+import com.hedera.services.evm.contracts.operations.HederaEvmStaticCallOperation;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.evm.EVM;
 import org.hyperledger.besu.evm.frame.MessageFrame;
 import org.hyperledger.besu.evm.gascalculator.GasCalculator;
 import org.hyperledger.besu.evm.operation.StaticCallOperation;
 import org.hyperledger.besu.evm.precompile.PrecompiledContract;
+
+import java.util.Map;
+import java.util.function.BiPredicate;
 
 /**
  * Hedera adapted version of the {@link StaticCallOperation}.
@@ -54,9 +57,8 @@ import org.hyperledger.besu.evm.precompile.PrecompiledContract;
  * transaction with {@link HederaExceptionalHaltReason#INVALID_SOLIDITY_ADDRESS} if the account does
  * not exist or it is deleted.
  */
-public class HederaStaticCallOperation extends StaticCallOperation {
+public class HederaStaticCallOperation extends HederaEvmStaticCallOperation {
     private final EvmSigsVerifier sigsVerifier;
-    private final BiPredicate<Address, MessageFrame> addressValidator;
     private final Map<String, PrecompiledContract> precompiledContractMap;
 
     public HederaStaticCallOperation(
@@ -64,9 +66,8 @@ public class HederaStaticCallOperation extends StaticCallOperation {
             final EvmSigsVerifier sigsVerifier,
             final BiPredicate<Address, MessageFrame> addressValidator,
             final Map<String, PrecompiledContract> precompiledContractMap) {
-        super(gasCalculator);
+        super(gasCalculator, addressValidator);
         this.sigsVerifier = sigsVerifier;
-        this.addressValidator = addressValidator;
         this.precompiledContractMap = precompiledContractMap;
     }
