+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/utils/ResponseCodeUtil.java
@@ -41,7 +41,7 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.*;
 import static java.util.stream.Collectors.toMap;
 
 import com.hedera.services.contracts.execution.TransactionProcessingResult;
-import com.hedera.services.contracts.operation.HederaExceptionalHaltReason;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.exceptions.ResourceLimitException;
 import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
 import java.util.Map;
