+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/contracts/operation/HederaCallCodeOperationTest.java
@@ -46,6 +46,7 @@ import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 
 import com.hedera.services.contracts.sources.EvmSigsVerifier;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.store.contracts.HederaStackedWorldStateUpdater;
 import java.util.Map;
 import java.util.Optional;
