+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/contracts/operation/HederaDelegateCallOperationTest.java
@@ -48,6 +48,9 @@ import static org.mockito.Mockito.lenient;
 
 import java.util.Optional;
 import java.util.function.BiPredicate;
+
+import com.hedera.services.evm.contracts.operations.HederaDelegateCallOperation;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import org.apache.tuweni.bytes.Bytes;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Wei;
