+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/contracts/operation/HederaExceptionalHaltReasonTest.java
@@ -17,6 +17,7 @@ package com.hedera.services.contracts.operation;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
