+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/contracts/operation/HederaExtCodeSizeOperationTest.java
@@ -23,6 +23,9 @@ import static org.mockito.BDDMockito.given;
 import java.util.Optional;
 import java.util.OptionalLong;
 import java.util.function.BiPredicate;
+
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
+import com.hedera.services.evm.contracts.operations.HederaExtCodeSizeOperation;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.evm.EVM;
