+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/contracts/operation/HederaOperationUtilTest.java
@@ -37,25 +37,13 @@ package com.hedera.services.contracts.operation;
  *
  */
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.never;
-import static org.mockito.BDDMockito.verify;
-
 import com.hedera.services.contracts.sources.EvmSigsVerifier;
+import com.hedera.services.evm.contracts.operations.HederaEvmOperationsUtil;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.store.contracts.HederaStackedWorldStateUpdater;
 import com.hedera.services.store.contracts.HederaWorldState;
 import com.hedera.services.store.contracts.WorldLedgers;
 import com.hedera.services.store.contracts.WorldStateAccount;
-import java.util.ArrayDeque;
-import java.util.Map;
-import java.util.Optional;
-import java.util.OptionalLong;
-import java.util.TreeMap;
-import java.util.function.LongSupplier;
-import java.util.function.Supplier;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
@@ -71,6 +59,21 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.ArrayDeque;
+import java.util.Map;
+import java.util.Optional;
+import java.util.OptionalLong;
+import java.util.TreeMap;
+import java.util.function.LongSupplier;
+import java.util.function.Supplier;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertSame;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.never;
+import static org.mockito.BDDMockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 class HederaOperationUtilTest {
@@ -119,7 +122,7 @@ class HederaOperationUtilTest {
 
         // when:
         final var result =
-                HederaOperationUtil.addressCheckExecution(
+                HederaEvmOperationsUtil.addressCheckExecution(
                         messageFrame,
                         () -> messageFrame.getStackItem(0),
                         gasSupplier,
@@ -144,7 +147,7 @@ class HederaOperationUtilTest {
 
         // when:
         final var result =
-                HederaOperationUtil.addressCheckExecution(
+                HederaEvmOperationsUtil.addressCheckExecution(
                         messageFrame,
                         () -> messageFrame.getStackItem(0),
                         gasSupplier,
@@ -172,7 +175,7 @@ class HederaOperationUtilTest {
 
         // when:
         final var result =
-                HederaOperationUtil.addressCheckExecution(
+                HederaEvmOperationsUtil.addressCheckExecution(
                         messageFrame,
                         () -> messageFrame.getStackItem(0),
                         gasSupplier,
