+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/contracts/operation/HederaStaticCallOperationTest.java
@@ -47,6 +47,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.lenient;
 
 import com.hedera.services.contracts.sources.EvmSigsVerifier;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.store.contracts.HederaStackedWorldStateUpdater;
 import java.util.Map;
 import java.util.Optional;
