+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/services/records/TransactionRecordServiceTest.java
@@ -17,9 +17,9 @@ package com.hedera.services.records;
 
 import static com.hedera.services.contracts.execution.traceability.CallOperationType.OP_CALL;
 import static com.hedera.services.contracts.execution.traceability.CallOperationType.OP_CREATE2;
-import static com.hedera.services.contracts.operation.HederaExceptionalHaltReason.*;
-import static com.hedera.services.contracts.operation.HederaExceptionalHaltReason.INVALID_SIGNATURE;
-import static com.hedera.services.contracts.operation.HederaExceptionalHaltReason.INVALID_SOLIDITY_ADDRESS;
+import static com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason.*;
+import static com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason.INVALID_SIGNATURE;
+import static com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason.INVALID_SOLIDITY_ADDRESS;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.*;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -32,7 +32,7 @@ import com.hedera.services.contracts.execution.TransactionProcessingResult;
 import com.hedera.services.contracts.execution.traceability.CallOperationType;
 import com.hedera.services.contracts.execution.traceability.ContractActionType;
 import com.hedera.services.contracts.execution.traceability.SolidityAction;
-import com.hedera.services.contracts.operation.HederaExceptionalHaltReason;
+import com.hedera.services.evm.contracts.operations.HederaExceptionalHaltReason;
 import com.hedera.services.ethereum.EthTxData;
 import com.hedera.services.state.submerkle.EntityId;
 import com.hedera.services.state.submerkle.EvmFnResult;
